backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
python-multipart
sqlalchemy
psycopg2-binary
pydantic-settings